+++ packages/core/src/presentation/index.ts
@@ -5,5 +5,6 @@
 
 export * from './labels.js'
 export * from './entity-emoji.js'
+export * from './tree-patterns.js'
 export * from './lenses.js'
 export * from './domain-rings.js'